requirements.txt
torch
torchvision
opencv-python-headless
streamlit
numpy
pandas
pillow
matplotlib
scikit-learn
